add virtualization for applications table
