+++ src/components/MonitoringApplicationsPage/MonitoringApplications/MonitoringApplications.jsx
@@ -19,70 +19,33 @@ such restriction.
 */
 import React, { useEffect, useMemo } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
-import { useNavigate, useParams } from 'react-router-dom'
+import { useParams } from 'react-router-dom'
 
-import ApplicationTableRow from '../../../elements/ApplicationTableRow/ApplicationTableRow'
 import MEPsWithDetections from './MEPsWithDetections'
 import NoData from '../../../common/NoData/NoData'
 import SectionTable from '../../../elements/SectionTable/SectionTable'
-import Table from '../../Table/Table'
-import { Loader, Tip } from 'igz-controls/components'
+import { Tip } from 'igz-controls/components'
+import AllApplicationsTable from './AllApplicationsTable'
 
-import { MODEL_ENDPOINTS_TAB, MONITORING_APP_PAGE } from '../../../constants'
 import { MONITORING_APPLICATIONS_NO_DATA_MESSAGE } from '../MonitoringApplicationsPage.util'
-import { createApplicationContent } from '../../../utils/createApplicationContent'
 import { generateOperatingFunctionsTable } from './monitoringApplications.util'
 import {
   removeMonitoringApplications,
   removeMEPWithDetections
 } from '../../../reducers/monitoringApplicationsReducer'
-import { saveAndTransformSearchParams } from 'igz-controls/utils/filter.util'
-
-import PresentMetricsIcon from 'igz-controls/images/present-metrics-icon.svg?react'
 
 const MonitoringApplications = () => {
   const dispatch = useDispatch()
   const params = useParams()
-  const navigate = useNavigate()
   const {
     monitoringApplications: { applications = [], operatingFunctions = [] },
     loading,
     error
   } = useSelector(store => store.monitoringApplicationsStore)
 
-  const applicationsTableActionsMenu = useMemo(
-    () => [
-      [],
-      [
-        {
-          id: 'open-metrics',
-          label: 'Open metrics',
-          icon: <PresentMetricsIcon />,
-          onClick: data =>
-            navigate(
-              `/projects/${params.projectName}/${MONITORING_APP_PAGE}/${data.name}/${MODEL_ENDPOINTS_TAB}${saveAndTransformSearchParams(
-                window.location.search,
-                true
-              )}`
-            )
-        }
-      ]
-    ],
-    [navigate, params.projectName]
-  )
   const operatingFunctionsTable = useMemo(
     () => generateOperatingFunctionsTable(operatingFunctions, params.projectName),
     [operatingFunctions, params.projectName]
-  )
-  const applicationsTableContent = useMemo(() => {
-    return applications.map(contentItem =>
-      createApplicationContent(contentItem, params.projectName)
-    )
-  }, [applications, params.projectName])
-
-  const applicationsTableHeaders = useMemo(
-    () => applicationsTableContent[0]?.content ?? [],
-    [applicationsTableContent]
   )
 
   useEffect(() => {
@@ -115,39 +78,7 @@ const MonitoringApplications = () => {
         </div>
       </div>
       <div className="monitoring-app__section section_big">
-        <div className="monitoring-app__section-item">
-          <div className="section-item_title">
-            <span>All applications</span>
-          </div>
-          {applications.length === 0 && !loading ? (
-            <NoData
-              message={
-                error
-                  ? 'Failed to fetch monitoring applications'
-                  : MONITORING_APPLICATIONS_NO_DATA_MESSAGE
-              }
-            />
-          ) : loading ? (
-            <Loader section secondary />
-          ) : (
-            <Table
-              actionsMenu={applicationsTableActionsMenu}
-              pageData={{}}
-              tableClassName="applications-table"
-              tableHeaders={applicationsTableHeaders}
-              skipTableWrapper
-            >
-              {applicationsTableContent.map((tableItem, index) => (
-                <ApplicationTableRow
-                  actionsMenu={applicationsTableActionsMenu}
-                  key={index}
-                  rowIndex={index}
-                  rowItem={tableItem}
-                />
-              ))}
-            </Table>
-          )}
-        </div>
+        <AllApplicationsTable applications={applications} loading={loading} error={error}/>
       </div>
     </div>
   )
